Refactored API folder into distinct modules.
