+++ src/api/Auth/README.md
@@ -1,0 +1,1 @@
+# Auth
